More Lambda tweaks:
We force LAMBDA_TASK_ROOT to our /app/lambda-handlers folder, to avoid
the complex logic of picking between the two.

It makes sense because the running_in_lambda() checks already use that
folder. If the script exists there, and then we switch to /var/task
folder where no handlers are found, it generates a weird runtime
error.